+++ src/engine.hpp
@@ -1,5 +1,5 @@
 #include "eval.hpp"
 #pragma once
-float alphabeta(Board board, float param[param_size], int depth, float alpha,
-                float beta);
-int go(Board board, float param[param_size], const Option &option);
+int alphabeta(Board board, int param[param_size], int depth, int alpha,
+                int beta);
+int go(Board board, int param[param_size], const Option &option);
