+++ src/play.cpp
@@ -1,7 +1,7 @@
 #include "play.hpp"
 
 // エンジン同士の対局
-int play_engine(float param_black[param_size], float param_white[param_size],
+int play_engine(int param_black[param_size], int param_white[param_size],
                 const Option &option) {
   Board board;
   int pass_count = 0;
